[refactor] one_button 다이얼로그화 #307
